+++ gm2m/query.py
@@ -3,8 +3,7 @@ from collections import defaultdict
 from django.db.models import query
 from django.utils import six
 
-from .contenttypes import ct as ct_classes
-from .helpers import get_content_type
+from .contenttypes import ct as ct_classes, get_content_type
 
 
 class GM2MTgtQuerySet(query.QuerySet):
@@ -51,7 +50,7 @@ class GM2MTgtQuerySet(query.QuerySet):
         for ct, attrs in six.iteritems(ct_attrs):
             for pk, obj in six.iteritems(
                 ct_classes.ContentType.objects.get_for_id(ct).model_class()
-                                     ._default_manager.in_bulk(attrs.keys())):
+                                      ._default_manager.in_bulk(attrs.keys())):
 
                 # we store the through model id in case we are in the process
                 # of fetching related objects